@@ -1,3 +1,3 @@
-module github.com/cavaliergopher/grab/v3
+module github.com/3JoB/grab/v3
 
-go 1.14
+go 1.20
