@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/cavaliergopher/grab/v3/pkg/grabui"
+	"github.com/3JoB/grab/v3/pkg/grabui"
 )
 
 func main() {
